Update README.md
This project is a "Contact Us" form designed to allow users to reach out with their contact information and messages. The form features a clean and responsive design, making it accessible and visually appealing on various devices. The project is implemented using HTML and CSS, ensuring simplicity and ease of customization.